requirements.txt
# Core dependencies
pywebview>=4.0
PyYAML>=6.0
GitPython>=3.1.0
Flask>=2.0.0

# Development dependencies (optional)
# pytest>=7.0
# pytest-cov>=4.0
# black>=22.0
# flake8>=5.0
# mypy>=1.0
